microservices/qrcode_service/requirements.txt
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
qrcode[pil]
Pillow==10.0.0
dj-database-url==1.3.0
django-environ==0.9.0
